Server: move out database connections, remove Redis expire time
